wpt_babel/content-security-policy/style-src/style-blocked.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="style-src 'none'" http-equiv="Content-Security-Policy"/>
  <title>
   style-blocked
  </title>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

async_test(function (t) {
  window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
    assert_equals(e.violatedDirective, "style-src-elem");
  }));
}, "Violated directive is script-src-elem.");
  </script>
  <link href="resources/blue.css" rel="stylesheet"/>
 </head>
 <body>
  <script>
   "use strict";

test(function (t) {
  assert_equals(document.styleSheets.length, 1);
}, "document.styleSheets should contain an item for the blocked CSS.");
  </script>
 </body>
</html>
